fix: copy of file if not set

+++ src/owa/entrypoint.sh
@@ -1,3 +1,8 @@
 #!/bin/sh
-cp /var/www/html/cfg/owa-config.php /var/www/html/owa-config.php
+sleep 5
+
+if [ -f /var/www/html/cfg/owa-config.php ]; then
+    cp /var/www/html/cfg/owa-config.php /var/www/html/owa-config.php
+fi
+
 exec apache2-foreground
